refatoracoes na croqui page e na sincronizacao

+++ lib/domain/services/achado_service.dart
@@ -13,9 +13,11 @@ class AchadoService {
 
   /// Registra uma nova lesão ou orifício anatômico ([Achado]) vinculado a um laudo.
   ///
-  /// @throws [Exception] se os dados obrigatórios do achado (como tipo de lesão ou coordenadas)
-  /// estiverem ausentes, ou se o laudo associado já estiver finalizado e assinado (bloqueado para edições).
+  /// @throws [Exception] se o laudo associado já estiver finalizado e assinado (bloqueado para edições).
   Future<void> salvarAchado(Achado achado) async {
+    if (await _repository.isCasoFinalizado(achado.casoUuid)) {
+      throw Exception('Segurança Jurídica: Este laudo já está finalizado e é imutável.');
+    }
     await _repository.insertAchado(achado);
   }
 
@@ -24,6 +26,9 @@ class AchadoService {
   /// @throws [Exception] se o laudo correspondente estiver finalizado, impedindo modificações
   /// retroativas sem a devida reabertura formal e auditoria do caso.
   Future<void> atualizarAchado(Achado achado) async {
+    if (await _repository.isCasoFinalizado(achado.casoUuid)) {
+      throw Exception('Segurança Jurídica: Este laudo já está finalizado e é imutável.');
+    }
     await _repository.updateAchado(achado);
   }
 
@@ -36,6 +41,10 @@ class AchadoService {
   ///
   /// @throws [Exception] se o laudo associado já estiver finalizado e assinado pelo [Perito].
   Future<void> removerAchado(String uuid) async {
+    final achado = await _repository.getAchadoByUuid(uuid);
+    if (achado != null && await _repository.isCasoFinalizado(achado.casoUuid)) {
+      throw Exception('Segurança Jurídica: Este laudo já está finalizado e é imutável.');
+    }
     await _repository.deleteAchado(uuid);
   }
 }
